+++ app/src/main/java/com/danosoftware/galaxyforce/sprites/game/splash/PlanetMovingSprite.java
@@ -24,8 +24,10 @@ public class PlanetMovingSprite extends AbstractMovingSprite {
 
     @Override
     public void animate(float deltaTime) {
-        timeElapsed += deltaTime;
-        int distance = (int) Math.min((timeElapsed - DELAY_IN_SECONDS_BEFORE_START) * DISTANCE_PER_SECOND, maxDistanceToTravel);
-        moveY(startPosition + distance);
+      timeElapsed += deltaTime;
+      float distance = Math
+          .min((timeElapsed - DELAY_IN_SECONDS_BEFORE_START) * DISTANCE_PER_SECOND,
+              maxDistanceToTravel);
+      moveY(startPosition + distance);
     }
 }
